refactor(goals)!: separate Goal and GoalDetail models
The `Goal` model is simplified for list views by removing the `createdAt`
property. A new `GoalDetail` model is introduced to represent the full
goal object, which is now returned by the creation endpoint.

The view model is updated to reload the entire list of goals after a
successful creation.

BREAKING CHANGE: The `GoalsRepository.create` method now returns a `GoalDetail` object instead of a `Goal`.